% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algo_CoOP_LBM.R
\name{best_ICL}
\alias{best_ICL}
\title{Best ICL}
\usage{
best_ICL(models)
}
\arguments{
\item{models}{a list of model with their ICL}
}
\value{
which model has the best ICL
}
\description{
Best ICL
}
\examples{

n1 = 100
n2 = 100
Q1=3
Q2=3
alpha1=c(.25,0.25, .5)
alpha2=c(0.10,0.4,0.5)
P <- matrix(c(0.9,0.6,0.4,0.7,0.5,0.3,0.5,0.3,0.1), Q1, Q2)
simulation1=simulate_lbm(P,alpha1,alpha2,n1,n2)
M = simulation1$A
Z1 =simulation1$Z
Z2 =simulation1$W
G= 300
lambda_i =rbeta(n1,0.3,1.5)
mu_j = rbeta(n2,0.3,1.5)
lambda_i = lambda_i/max(lambda_i)
mu_j = mu_j/max(mu_j)
N0=lambda_i\%*\%t(mu_j)
N0 = N0*G
N=matrix(rpois(n1*n2,N0),nrow=n1)
R = M*N
obsrow = rowSums(R)>0
obscol = colSums(R)>0
R_obs = R[obsrow,obscol]
M_obs = M[obsrow,obscol]
R = M*N
models=main_CoOP_LBM(R)
print(best_model(models))
}
