% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_boundaries}
\alias{check_boundaries}
\title{Check boundaries}
\usage{
check_boundaries(x, zero = .Machine$double.eps)
}
\arguments{
\item{x}{numeric vector or matrix}

\item{zero}{smallest tolerance allowed}
}
\value{
round element too close of zero or one to a value in ]0,1[
}
\description{
Check boundaries
}
\examples{
a<- c(0.2,0.5,0)
log(a)
log(check_boundaries(a))
}
