% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clustering_indicator}
\alias{clustering_indicator}
\title{Clustering indicator}
\usage{
clustering_indicator(clustering, k)
}
\arguments{
\item{clustering}{vector of labels}

\item{k}{integer of maximum number of labels}
}
\value{
clustering indicator with k column
}
\description{
Clustering indicator
}
\examples{
a <- c(1,1,1,2,2,2,3,3)
clustering_indicator(a,3)
clustering_indicator(a,4)
}
