% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clustinit_LBM}
\alias{clustinit_LBM}
\title{Clustering initialization}
\usage{
clustinit_LBM(connectivity, Q1, Q2, type = "hierarchical_clust")
}
\arguments{
\item{connectivity}{Binary matrix}

\item{Q1}{number of clusters for rows}

\item{Q2}{number of clusters for columns}

\item{type}{type of initialization : "hierarchical_clust", "spectral_clust" or "kmeans_clust"}
}
\value{
list containing an initial clustering for rows and columns
}
\description{
Clustering initialization
}
\examples{
a<- matrix(0,10,10)
a[1:5,1:5] = runif(25)<0.9
a[6:10,1:5] = runif(25)<0.5
a[1:5,6:10] = runif(25)<0.3
a[6:10,6:10] = runif(25)<0.1
print(a)
print(clustinit_LBM(a,2,2))
}
