% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algo_CoOP_LBM.R
\name{CoOP_LBM}
\alias{CoOP_LBM}
\title{Expected Degree Distribution and Variational Expectation Maximization}
\usage{
CoOP_LBM(R, Q1, Q2, Z1 = c(), Z2 = c(), estimOptions = list())
}
\arguments{
\item{R}{observed connectivity matrix}

\item{Q1}{number of row clusters}

\item{Q2}{number of columns clusters}

\item{tau1}{initial clustering of rows}

\item{tau2}{initial clustering of columns}

\item{param}{list of parameters}
}
\value{
Estimated LBM parameters, clustering, lambda and mu
}
\description{
Expected Degree Distribution and Variational Expectation Maximization
}
