% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOPLBM-package.R
\docType{package}
\name{CoOPLBM-package}
\alias{CoOPLBM}
\alias{CoOPLBM-package}
\title{CoOPLBM: Corrected Observation Process for Latent Block Model}
\description{
By taking into account the counting data, it is possible to correct a latent block model (LBM) where data could have been missing. This package allows to fit a CoOP-LBM on a bipartite graph, represented as a contigency table with the number of interaction observed.
}
\author{
\strong{Maintainer}: Emre ANAKOK \email{emre.anakok@agroparistech.fr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
