% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.softmax}
\alias{.softmax}
\title{Softmax}
\usage{
.softmax(x)
}
\arguments{
\item{x}{numeric vector or matrix}
}
\value{
softmax function applied to the vector or matrix
}
\description{
Softmax
}
\examples{
a<- c(10,20,30)
.softmax(a)
}
