% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algo_CoOP_LBM.R
\name{forward_explo}
\alias{forward_explo}
\title{forward exploration}
\usage{
forward_explo(models, k1, k2, connectivity, f, param)
}
\arguments{
\item{models}{list of models}

\item{k1}{number of row clusters}

\item{k2}{number of column clusters}

\item{connectivity}{connectivity matrix}

\item{f}{method (VEM or EDD-VEM)}

\item{param}{list of parameters}
}
\value{
forward exploration of a given list of model
}
\description{
forward exploration
}
