% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LBM_connectivity_prob3}
\alias{LBM_connectivity_prob3}
\title{Update connectivity}
\usage{
LBM_connectivity_prob3(V, Z1, Z2, pi, lambda_mu)
}
\arguments{
\item{V}{binary connectivity matrix}

\item{Z1}{clustering of rows}

\item{Z2}{clustering of columns}

\item{pi}{matrix of probabilities of connections between groups}

\item{lambda_mu}{matrix of all the coefficient lambda x mu x G}
}
\value{
probability of M = 1 given that V = 0. Only return the values where V = 0
}
\description{
Update connectivity
}
