% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LBM_ICL}
\alias{LBM_ICL}
\title{ICL for LBM}
\usage{
LBM_ICL(members1, members2, alpha1, alpha2, pi, connectivity)
}
\arguments{
\item{members1}{clustering indicator of rows}

\item{members2}{clustering indicator of column}

\item{alpha1}{proportion of group by row}

\item{alpha2}{proportion of group by column}

\item{pi}{matrix of probabilities of connections between groups}

\item{connectivity}{binary connectivity matrix}
}
\value{
estimation of ICL for classic LBM
}
\description{
ICL for LBM
}
