% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LBM_ICL_3}
\alias{LBM_ICL_3}
\title{ICL for CoOP-LBM (Poisson)}
\usage{
LBM_ICL_3(members1, members2, alpha1, alpha2, pi, R, lambdamu)
}
\arguments{
\item{members1}{clustering indicator of rows}

\item{members2}{clustering indicator of column}

\item{alpha1}{proportion of group by row}

\item{alpha2}{proportion of group by column}

\item{pi}{matrix of probabilities of connections between groups}

\item{R}{observed connectivity matrix}

\item{lambdamu}{matrix of lambda_i x lambda_j}
}
\value{
estimation of ICL for EDD LBM using the likelihood taking into account the Poisson distribution
}
\description{
ICL for CoOP-LBM (Poisson)
}
