% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LBM_update_alpha}
\alias{LBM_update_alpha}
\title{Update alpha for VEM}
\usage{
LBM_update_alpha(tau)
}
\arguments{
\item{tau}{soft clustering (row or column)}
}
\value{
new estimation of alpha (row or column)
}
\description{
Update alpha for VEM
}
