% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LBM_update_alpha3}
\alias{LBM_update_alpha3}
\title{Update alpha for Gibbs}
\usage{
LBM_update_alpha3(Z, Q)
}
\arguments{
\item{Z}{clustering (row or column)}

\item{Q}{number of group (row or column)}
}
\value{
new estimation of alpha or beta
}
\description{
Update alpha for Gibbs
}
