% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LBM_update_connectivity3}
\alias{LBM_update_connectivity3}
\title{Update connectivity}
\usage{
LBM_update_connectivity3(V, Z1, Z2, pi, lambda_mu)
}
\arguments{
\item{V}{binary connectivity matrix of the observed matrix}

\item{Z1}{clustering by row}

\item{Z2}{clustering by column}

\item{pi}{probability of connection between groups}

\item{lambda_mu}{matrix containing the product lambda_i x mu_j x G}
}
\value{
simulation of a connectivity matrix with missing links added
}
\description{
Update connectivity
}
