% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LBM_update_lambda_mu}
\alias{LBM_update_lambda_mu}
\title{Update lambda and mu}
\usage{
LBM_update_lambda_mu(rowSumsR, colSumsR, connectivity, fixPointIter = 3)
}
\arguments{
\item{rowSumsR}{sum of observation by row}

\item{colSumsR}{sum of observation by column}

\item{connectivity}{binary connectivity matrix}

\item{fixPointIter}{number of iteration for the fixed point algorithm}
}
\value{
list containing the estimation of lambda_i, mu_j x G, and the matrix lambda_i x mu_j x G
}
\description{
Update lambda and mu
}
