% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LBM_update_pi}
\alias{LBM_update_pi}
\title{Update pi for VEM}
\usage{
LBM_update_pi(connectivity, tau1, tau2)
}
\arguments{
\item{connectivity}{binary matrix of connectivity}

\item{tau1}{soft clustering of rows}

\item{tau2}{soft clustering of columns}
}
\value{
new estimation of pi
}
\description{
Update pi for VEM
}
