% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LBM_update_pi3}
\alias{LBM_update_pi3}
\title{Update pi for Gibbs}
\usage{
LBM_update_pi3(connectivity, Z1, Z2, Q1, Q2)
}
\arguments{
\item{connectivity}{binary connectivity matrix}

\item{Z1}{row clustering}

\item{Z2}{column clustering}

\item{Q1}{number of groups for rows}

\item{Q2}{number of groups for columns}
}
\value{
new value of pi given the parameters
}
\description{
Update pi for Gibbs
}
