% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LBM_update_tau}
\alias{LBM_update_tau}
\title{Update tau for VEM}
\usage{
LBM_update_tau(connectivity, alpha1, alpha2, pi, tau1, tau2)
}
\arguments{
\item{connectivity}{binary matrix of connectivity}

\item{alpha1}{proportion of groups by row}

\item{alpha2}{proportion of groups by column}

\item{pi}{matrix of probabilities of connections between groups}

\item{tau1}{soft clustering of rows}

\item{tau2}{soft clustering of columns}
}
\value{
new estimation of tau
}
\description{
Update tau for VEM
}
