% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LBM_update_Z}
\alias{LBM_update_Z}
\title{Update clustering}
\usage{
LBM_update_Z(R, alpha1, alpha2, pi, lambda_mu, Z1, Z2, lfactorialR)
}
\arguments{
\item{R}{matrix of observation}

\item{alpha1}{proportion of clustering by row}

\item{alpha2}{proportion of clustering by column}

\item{pi}{matrix of probabilities of connections between groups}

\item{lambda_mu}{matrix containing the product lambda_i x mu_j x G}

\item{Z1}{clustering by row}

\item{Z2}{clustering by column}

\item{lfactorialR}{matrix of log factorial of the R matrix}
}
\value{
simulate new values for Z1 and Z2 given the parameters
}
\description{
Update clustering
}
