% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algo_CoOP_LBM.R
\name{LBM_VEM}
\alias{LBM_VEM}
\title{Variational Expectation Maximization for LBM}
\usage{
LBM_VEM(connectivity, Q1, Q2, Z1 = c(), Z2 = c(), estimOptions = list())
}
\arguments{
\item{connectivity}{binary matrix of connectivity}

\item{Q1}{number of row clusters}

\item{Q2}{number of column clusters}

\item{Z1}{initial clustering of rows}

\item{Z2}{initial clustering of columns}

\item{param}{list of parameters}
}
\value{
Estimated LBM parameters and clustering
}
\description{
Variational Expectation Maximization for LBM
}
