% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{memberships}
\alias{memberships}
\title{Memberships}
\usage{
memberships(tau)
}
\arguments{
\item{tau}{}
}
\value{
hard clustering indicator of a soft clustering
}
\description{
Memberships
}
\examples{
a<- matrix(0,5,2)
a[,1] = runif(5)
a[,2] = 1-a[,1]
print(a)
print(memberships(a))
}
