% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{membertoclust}
\alias{membertoclust}
\title{Membership to clustering}
\usage{
membertoclust(membership)
}
\arguments{
\item{membership}{a clustering indicator matrix}
}
\value{
the clustering associated to the matrix
}
\description{
Membership to clustering
}
