% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quad_form}
\alias{quad_form}
\title{Quadratic form}
\usage{
quad_form(A, x)
}
\arguments{
\item{A}{square matrix}

\item{x}{vector}
}
\value{
quadratic form associated to the matrix A applied in x
}
\description{
Quadratic form
}
