% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoOPLBM-package.R
\docType{package}
\name{CoOPLBM-package}
\alias{CoOPLBM}
\alias{CoOPLBM-package}
\title{CoOPLBM: Corrected Observation Process for Latent Block Model}
\description{
Fits a LBM on a weigthed bipartite graph taking into account that aunobserved data can be missing data.
}
\author{
\strong{Maintainer}: Emre Anakok \email{emre.anakok@agroparistech.fr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
