% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algo_CoOP_LBM.R
\name{fit_supervised_CoOP_LBM}
\alias{fit_supervised_CoOP_LBM}
\title{Corrected Observation Process for Latent Block Model}
\usage{
fit_supervised_CoOP_LBM(R, Q1, Q2, Z1 = c(), Z2 = c(), estimOptions = list())
}
\arguments{
\item{R}{observed connectivity matrix}

\item{Q1}{number of row clusters}

\item{Q2}{number of columns clusters}

\item{Z1}{initial clustering of rows}

\item{Z2}{initial clustering of columns}

\item{estimOptions}{list of parameters}
}
\value{
Estimated LBM parameters, clustering, lambda, mu and G for a given number of groups.
}
\description{
Corrected Observation Process for Latent Block Model
}
