% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algo_CoOP_LBM.R
\name{fit_supervised_LBM}
\alias{fit_supervised_LBM}
\title{Variational Expectation Maximization for LBM}
\usage{
fit_supervised_LBM(
  connectivity,
  Q1,
  Q2,
  Z1 = c(),
  Z2 = c(),
  estimOptions = list()
)
}
\arguments{
\item{connectivity}{Binary matrix of connectivity}

\item{Q1}{Number of row clusters}

\item{Q2}{Number of column clusters}

\item{Z1}{Initial clustering of rows}

\item{Z2}{Initial clustering of columns}

\item{param}{List of parameters}
}
\value{
Estimated LBM parameters and clustering
}
\description{
Variational Expectation Maximization for LBM
}
