% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NODF}
\alias{NODF}
\title{Nestedness metric based on Overlap and Decreasing Fill}
\usage{
NODF(Matrix)
}
\arguments{
\item{Matrix}{}
}
\value{
List of NODF of the matrix,averaged by row, by column, and for the full matrix
}
\description{
Nestedness metric based on Overlap and Decreasing Fill
}
